Ensure state changes happen in order and flagged values are not overwritten
It's possible to encounter race conditions when updating the panelItems and menuItems
if you have a ToolsPanelItem that's displayed conditionally, or which is shown by
default only conditionally. This can cause multiple calls to update the state in quick
sequence, which can result in those calls overwriting one another. To prevent this,
the calls to setState have been updated to depend on the previous state.